customer/requirements.txt
Flask
Redis
kafka-python

